refactor state management hook into separate file
